set model weights to bfloat16

+++ app/core/engine.py
@@ -22,7 +22,7 @@ class DeepSeekOCR:
             trust_remote_code=True,
             use_safetensors=True,
         )
-        self.model.eval().to("cuda")
+        self.model.eval().to(torch.bfloat16).to("cuda")
         self.prompt = "<image>\n<|grounding|>Convert the document to markdown."
 
     def infer(
